docs: Update base token section with deprecated
