+++ packages/canvas-tokens-docs/.storybook/global.css
@@ -78,7 +78,6 @@ h5,
 h6 {
   display: flex;
   align-items: center;
-  gap: 8px;
 }
 
 pre {
