+++ packages/canvas-tokens-docs/components/StatusIndicator/index.css
@@ -4,6 +4,7 @@
   gap: 8px;
   padding: 0 8px;
   border-radius: var(--cnvs-sys-shape-round);
+  margin-inline-start: 8px;
 }
 
 .status-indicator-icon {
