+++ packages/canvas-tokens-docs/stories/base/Base.stories.mdx
@@ -1,15 +1,9 @@
 import {Meta, Unstyled} from '@storybook/blocks';
 
-import {
-  AIColorTokens,
-  PrimaryColorTokens,
-  NeutralColorTokens,
-  TypeColorTokens,
-  StatusColorTokens,
-  OtherColorTokens,
-  DeprecatedColorTokens,
-} from './examples/Color';
+import {ColorTokens, DeprecatedColorTokens} from './examples/Color';
+import {DeprecatedTokens} from './examples/DeprecatedTokens';
 import {BaseUnit} from './examples/BaseUnit';
+import {NavCard} from '../../components/NavCard';
 
 <Meta title="Docs/Base Tokens" />
 
@@ -73,73 +67,26 @@ UI.
 
 ## Color
 
-Base colors provide the core palettes that make up our UI. There are twenty-seven palettes in total,
-but only a handful are used regularly. These palettes don't have any inherent contextual meaning on
-when and where they should be used. However, to provide some guidance, we've grouped them
-semantically: primary, neutral, type, status, and other.
+Base colors provide the core palettes that make up our UI. There are thirteen palettes in total and
+they don't have any inherent contextual meaning on when and where they should be used.
 
-### Primary
-
-Blueberry is our primary color and the most prominent palette in our UI. We use it for our primary
-buttons, links, focus outlines, and more.
-
-<PrimaryColorTokens />
-
----
-
-### Neutral
-
-We have four neutral palettes: french vanilla, soap, licorice, and black pepper. These colors are
-used for our backgrounds, foregrounds, icons, and text.
-
-<NeutralColorTokens />
-
----
-
-### Type
-
-Our type colors are used for headings, text, and icons. Maintaining sufficient contrast is an
-important consideration when choosing any type color. Consider where your type will render as well
-as any focus, active, and hover states.
-
-<TypeColorTokens />
-
----
-
-### Status
-
-These palettes are reserved to indicate the status of a particular object, task, or flow. These
-palettes are generally better suited for backgrounds than text, as they have difficulty meeting our
-contrast ratios guidelines. Also remember to always communicate status with more than color alone.
-
-<StatusColorTokens />
-
----
-
-### AI
-
-This palette is reserved to indicate the use of artificial intelligence, such as loading or
-generating content.
-
-<AIColorTokens />
-
----
-
-### Other
-
-Our other palettes have little-to-no usage in our UI outside of data visualization. Please, use
-these sparingly.
-
-<OtherColorTokens />
+<br />
+<ColorTokens />
 
 ---
 
 ### Deprecated
 
-The following fruit based palettes are deprecated and will be removed in a future release. Please use the new color palette. While the base tokens are deprecated, we advise to use system tokens instead for better future proofing. 
+The fruit based palettes are deprecated and will be removed in a future release. Please use the new
+color palette. While the base tokens are deprecated, we advise to use system tokens instead for
+better future proofing.
 
-For more information on how to migrate, please view our [Token Migration guide](https://workday.github.io/canvas-kit/?path=/docs/guides-tokens-migration-overview--docs).
+<NavCard
+  title="Deprecated Base Tokens"
+  link="/?path=/docs/guides-upgrade-guides-v3-visual-changes-base-deprecated-palette--docs"
+/>
 
-<DeprecatedColorTokens />
+For more information on how to migrate, please view our
+[Token Migration guide](https://workday.github.io/canvas-tokens/?path=/docs/guides-upgrade-guides-v3-overview--docs).
 
 </Unstyled>
