+++ packages/canvas-tokens-docs/stories/base/examples/DeprecatedTokens.tsx
@@ -1,31 +1,28 @@
 import * as React from 'react';
-import {updatedSystemTokens} from './data/deprecatedMapping';
+import {deprecatedBaseTokens} from './data/deprecatedMapping';
 import {TokenGrid} from '../../../components/TokenGrid';
 import {SwatchWithText} from '../../../components/SwatchWithText';
 
-export const BorderUpdates = () => {
-  const {border} = updatedSystemTokens;
-  const names = Object.keys(border);
+export const DeprecatedTokens = () => {
+  const names = Object.keys(deprecatedBaseTokens);
 
   return (
     <div style={{display: 'flex', flexDirection: 'column', gap: '2rem', marginTop: '2rem'}}>
       {names.map(name => (
         <TokenGrid
-          caption={`${name} border tokens`.toUpperCase()}
-          headings={['Token Name', 'Old Value', 'New Value', 'Change']}
-          rows={border[name]}
+          caption={name.replaceAll('-', ' ')}
+          headings={['Token Name', 'Old Value', 'New Value', 'System Token Replacement']}
+          rows={deprecatedBaseTokens[name]}
         >
           {(token: any) => (
             <>
               <TokenGrid.RowItem>{token.name}</TokenGrid.RowItem>
-              <SwatchWithText
-                color={token.oldColor}
-                name={token.old}
-                withWhiteText={token.withWhiteText}
-              />
+              <SwatchWithText color={token.old} withWhiteText={token.withWhiteText} />
               <SwatchWithText color={token.new} withWhiteText={token.withWhiteText} />
               <TokenGrid.RowItem>
-                {!token.old ? '➕ Added' : !token.new ? '❌ Removed' : '✅ Updated'}
+                {token.systemTokens?.map(token => (
+                  <div>{token}</div>
+                ))}
               </TokenGrid.RowItem>
             </>
           )}
