+++ packages/canvas-tokens-docs/stories/base/examples/data/deprecatedMapping.ts
@@ -1,0 +1,1032 @@
+export const deprecatedBaseTokens = {
+  cinnamon: [
+    {
+      name: 'base.palette.cinnamon.100',
+      old: '#ffefee',
+      new: 'base.palette.red.50',
+      systemTokens: ['sys.color.bg.critical.softer'],
+    },
+    {
+      name: 'base.palette.cinnamon.200',
+      old: '#FCC9C5',
+      new: 'base.palette.red.100',
+      systemTokens: ['sys.color.bg.critical.soft'],
+    },
+    {
+      name: 'base.palette.cinnamon.300',
+      old: '#ff867d',
+      new: 'base.palette.red.300',
+    },
+    {
+      name: 'base.palette.cinnamon.400',
+      old: '#ff5347',
+      new: 'base.palette.red.400',
+      systemTokens: ['sys.color.fg.critical.soft'],
+    },
+    {
+      name: 'base.palette.cinnamon.500',
+      old: '#de2e21',
+      new: 'base.palette.red.600',
+      systemTokens: [
+        'sys.color.bg.critical.default',
+        'sys.color.fg.critical.default',
+        'sys.color.border.critical.default',
+      ],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.cinnamon.600',
+      old: '#a31b12',
+      new: 'base.palette.red.700',
+      systemTokens: ['sys.color.bg.critical.strong', 'sys.color.fg.critical.strong'],
+      withWhiteText: true,
+    },
+  ],
+  peach: [
+    {
+      name: 'base.palette.peach.100',
+      old: '#fff3f0',
+      new: 'base.palette.coral.50',
+    },
+    {
+      name: 'base.palette.peach.200',
+      old: '#ffc2b3',
+      new: 'base.palette.coral.200',
+    },
+    {
+      name: 'base.palette.peach.300',
+      old: '#ff957a',
+      new: 'base.palette.coral.300',
+    },
+    {
+      name: 'base.palette.peach.400',
+      old: '#ff643d',
+      new: 'base.palette.red.400',
+    },
+    {
+      name: 'base.palette.peach.500',
+      old: '#de4721',
+      new: 'base.palette.coral.600',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.peach.600',
+      old: '#b53413',
+      new: 'base.palette.coral.700',
+      withWhiteText: true,
+    },
+  ],
+  'chili-mango': [
+    {
+      name: 'base.palette.chili-mango.100',
+      old: '#ffe6d9',
+      new: 'base.palette.coral.25',
+    },
+    {
+      name: 'base.palette.chili-mango.200',
+      old: '#ffc7ab',
+      new: 'base.palette.coral.200',
+    },
+    {
+      name: 'base.palette.chili-mango.300',
+      old: '#ff9b69',
+      new: 'base.palette.coral.300',
+    },
+    {
+      name: 'base.palette.chili-mango.400',
+      old: '#ff671b',
+      new: 'base.palette.orange.500',
+    },
+    {
+      name: 'base.palette.chili-mango.500',
+      old: '#e04b00',
+      new: 'base.palette.orange.500',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.chili-mango.600',
+      old: '#a33600',
+      new: 'base.palette.orange.700',
+      withWhiteText: true,
+    },
+  ],
+  cantaloupe: [
+    {
+      name: 'base.palette.cantaloupe.100',
+      old: '#ffeed9',
+      new: 'base.palette.amber.50',
+      systemTokens: ['sys.color.bg.caution.softer'],
+    },
+    {
+      name: 'base.palette.cantaloupe.200',
+      old: '#fcd49f',
+      new: 'base.palette.amber.200',
+      systemTokens: ['sys.color.fg.caution.softer'],
+    },
+    {
+      name: 'base.palette.cantaloupe.300',
+      old: '#ffbc63',
+      new: 'base.palette.amber.300',
+      systemTokens: ['sys.color.bg.caution.soft'],
+    },
+    {
+      name: 'base.palette.cantaloupe.400',
+      old: '#ffa126',
+      new: 'base.palette.amber.400',
+      systemTokens: ['sys.color.bg.caution.default', 'sys.color.border.caution.default'],
+    },
+    {
+      name: 'base.palette.cantaloupe.500',
+      old: '#f38b00',
+      new: 'base.palette.amber.500',
+      systemTokens: ['sys.color.bg.caution.strong'],
+    },
+    {
+      name: 'base.palette.cantaloupe.600',
+      old: '#c06c00',
+      new: 'base.palette.amber.600',
+      systemTokens: ['sys.color.bg.caution.stronger', 'sys.color.border.caution.strong'],
+      withWhiteText: true,
+    },
+  ],
+  'sour-lemon': [
+    {
+      name: 'base.palette.sour-lemon.100',
+      old: '#fff9e6',
+      new: 'base.palette.amber.25',
+    },
+    {
+      name: 'base.palette.sour-lemon.200',
+      old: '#ffecab',
+      new: 'base.palette.amber.100',
+    },
+    {
+      name: 'base.palette.sour-lemon.300',
+      old: '#ffda61',
+      new: 'base.palette.amber.200',
+    },
+    {
+      name: 'base.palette.sour-lemon.400',
+      old: '#ffc629',
+      new: 'base.palette.amber.300',
+    },
+    {
+      name: 'base.palette.sour-lemon.500',
+      old: '#ebb400',
+      new: 'base.palette.amber.300',
+    },
+    {
+      name: 'base.palette.sour-lemon.600',
+      old: '#bd9100',
+      new: 'base.palette.amber.500',
+      withWhiteText: true,
+    },
+  ],
+  'juicy-pear': [
+    {
+      name: 'base.palette.juicy-pear.100',
+      old: '#f7fae6',
+      new: 'base.palette.amber.25',
+    },
+    {
+      name: 'base.palette.juicy-pear.200',
+      old: '#e2f391',
+      new: 'base.palette.amber.100',
+    },
+    {
+      name: 'base.palette.juicy-pear.300',
+      old: '#c4de40',
+      new: 'base.palette.amber.200',
+    },
+    {
+      name: 'base.palette.juicy-pear.400',
+      old: '#a8c224',
+      new: 'base.palette.amber.200',
+    },
+    {
+      name: 'base.palette.juicy-pear.500',
+      old: '#8ea618',
+      new: 'base.palette.green.500',
+    },
+    {
+      name: 'base.palette.juicy-pear.600',
+      old: '#687818',
+      new: 'base.palette.green.700',
+      withWhiteText: true,
+    },
+  ],
+  kiwi: [
+    {
+      name: 'base.palette.kiwi.100',
+      old: '#ecfcd7',
+      new: 'base.palette.green.50',
+    },
+    {
+      name: 'base.palette.kiwi.200',
+      old: '#caf593',
+      new: 'base.palette.green.100',
+    },
+    {
+      name: 'base.palette.kiwi.300',
+      old: '#a7e05c',
+      new: 'base.palette.green.200',
+    },
+    {
+      name: 'base.palette.kiwi.400',
+      old: '#77bc1f',
+      new: 'base.palette.green.500',
+    },
+    {
+      name: 'base.palette.kiwi.500',
+      old: '#609915',
+      new: 'base.palette.green.500',
+    },
+    {
+      name: 'base.palette.kiwi.600',
+      old: '#537824',
+      new: 'base.palette.green.700',
+      withWhiteText: true,
+    },
+  ],
+  'green-apple': [
+    {
+      name: 'base.palette.green-apple.100',
+      old: '#ebfff0',
+      new: 'base.palette.green.50',
+      systemTokens: ['sys.color.bg.positive.softer'],
+    },
+    {
+      name: 'base.palette.green-apple.200',
+      old: '#acf5be',
+      new: 'base.palette.green.100',
+      systemTokens: ['sys.color.bg.positive.soft'],
+    },
+    {
+      name: 'base.palette.green-apple.300',
+      old: '#5fe380',
+      new: 'base.palette.green.200',
+      systemTokens: ['sys.color.fg.positive.soft'],
+    },
+    {
+      name: 'base.palette.green-apple.400',
+      old: '#43c463',
+      new: 'base.palette.green.600',
+      systemTokens: ['sys.color.bg.positive.default', 'sys.color.fg.positive.default'],
+    },
+    {
+      name: 'base.palette.green-apple.500',
+      old: '#319c4c',
+      new: 'base.palette.green.700',
+      systemTokens: ['sys.color.bg.positive.strong', 'sys.color.fg.positive.strong'],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.green-apple.600',
+      old: '#217a37',
+      new: 'base.palette.green.800',
+      systemTokens: ['sys.color.bg.positive.stronger', 'sys.color.fg.positive.stronger'],
+      withWhiteText: true,
+    },
+  ],
+  watermelon: [
+    {
+      name: 'base.palette.watermelon.100',
+      old: '#ebfdf8',
+      new: 'base.palette.neutral.50',
+    },
+    {
+      name: 'base.palette.watermelon.200',
+      old: '#b7edde',
+      new: 'base.palette.neutral.100',
+    },
+    {
+      name: 'base.palette.watermelon.300',
+      old: '#65ccaf',
+      new: 'base.palette.green.100',
+    },
+    {
+      name: 'base.palette.watermelon.400',
+      old: '#12a67c',
+      new: 'base.palette.green.600',
+    },
+    {
+      name: 'base.palette.watermelon.500',
+      old: '#0b7a5c',
+      new: 'base.palette.green.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.watermelon.600',
+      old: '#08513d',
+      new: 'base.palette.green.900',
+      withWhiteText: true,
+    },
+  ],
+  jewel: [
+    {
+      name: 'base.palette.jewel.100',
+      old: '#ebfdff',
+      new: 'base.palette.teal.25',
+    },
+    {
+      name: 'base.palette.jewel.200',
+      old: '#acecf3',
+      new: 'base.palette.teal.200',
+    },
+    {
+      name: 'base.palette.jewel.300',
+      old: '#44c8d7',
+      new: 'base.palette.teal.400',
+    },
+    {
+      name: 'base.palette.jewel.400',
+      old: '#1ea4b3',
+      new: 'base.palette.teal.500',
+    },
+    {
+      name: 'base.palette.jewel.500',
+      old: '#1a818c',
+      new: 'base.palette.teal.600',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.jewel.600',
+      old: '#156973',
+      new: 'base.palette.teal.700',
+      withWhiteText: true,
+    },
+  ],
+  toothpaste: [
+    {
+      name: 'base.palette.toothpaste.100',
+      old: '#d7f1fc',
+      new: 'base.palette.azure.50',
+    },
+    {
+      name: 'base.palette.toothpaste.200',
+      old: '#99e0ff',
+      new: 'base.palette.azure.200',
+    },
+    {
+      name: 'base.palette.toothpaste.300',
+      old: '#40b4e5',
+      new: 'base.palette.azure.300',
+    },
+    {
+      name: 'base.palette.toothpaste.400',
+      old: '#1894c9',
+      new: 'base.palette.azure.400',
+    },
+    {
+      name: 'base.palette.toothpaste.500',
+      old: '#0271a1',
+      new: 'base.palette.azure.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.toothpaste.600',
+      old: '#005B82',
+      new: 'base.palette.azure.800',
+      withWhiteText: true,
+    },
+  ],
+  blueberry: [
+    {
+      name: 'base.palette.blueberry.100',
+      old: '#D7EAFC',
+      new: 'base.palette.blue.25',
+    },
+    {
+      name: 'base.palette.blueberry.200',
+      old: '#A6D2FF',
+      new: 'base.palette.blue.100',
+      systemTokens: ['sys.color.bg.primary.soft'],
+    },
+    {
+      name: 'base.palette.blueberry.300',
+      old: '#40A0FF',
+      new: 'base.palette.blue.400',
+      systemTokens: ['sys.color.bg.primary.default', 'sys.color.fg.primary.soft'],
+    },
+    {
+      name: 'base.palette.blueberry.400',
+      old: '#0875E1',
+      new: 'base.palette.blue.600',
+      systemTokens: [
+        'sys.color.bg.primary.default',
+        'sys.color.fg.primary.default',
+        'sys.color.border.primary.default',
+      ],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.blueberry.500',
+      old: '#005cb9',
+      new: 'base.palette.blue.700',
+      systemTokens: ['sys.color.bg.primary.strong', 'sys.color.fg.primary.strong'],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.blueberry.600',
+      old: '#004387',
+      new: 'base.palette.blue.800',
+      systemTokens: ['sys.color.bg.primary.stronger', 'sys.color.text.primary.stronger'],
+      withWhiteText: true,
+    },
+  ],
+  plum: [
+    {
+      name: 'base.palette.plum.100',
+      old: '#e6f1ff',
+      new: 'base.palette.blue.100',
+    },
+    {
+      name: 'base.palette.plum.200',
+      old: '#A6CDFF',
+      new: 'base.palette.blue.200',
+    },
+    {
+      name: 'base.palette.plum.300',
+      old: '#529bfa',
+      new: 'base.palette.blue.400',
+    },
+    {
+      name: 'base.palette.plum.400',
+      old: '#3881E1',
+      new: 'base.palette.blue.600',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.plum.500',
+      old: '#3166ab',
+      new: 'base.palette.blue.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.plum.600',
+      old: '#264a7a',
+      new: 'base.palette.blue.800',
+      withWhiteText: true,
+    },
+  ],
+  'berry-smoothie': [
+    {
+      name: 'base.palette.berry-smoothie.100',
+      old: '#e8edff',
+      new: 'base.palette.indigo.50',
+    },
+    {
+      name: 'base.palette.berry-smoothie.200',
+      old: '#c2cfff',
+      new: 'base.palette.indigo.200',
+    },
+    {
+      name: 'base.palette.berry-smoothie.300',
+      old: '#7891FF',
+      new: 'base.palette.indigo.400',
+    },
+    {
+      name: 'base.palette.berry-smoothie.400',
+      old: '#5E77E6',
+      new: 'base.palette.blue.500',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.berry-smoothie.500',
+      old: '#4b5eb3',
+      new: 'base.palette.blue.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.berry-smoothie.600',
+      old: '#3b4987',
+      new: 'base.palette.blue.800',
+      withWhiteText: true,
+    },
+  ],
+  blackberry: [
+    {
+      name: 'base.palette.blackberry.100',
+      old: '#f0f0ff',
+      new: 'base.palette.indigo.25',
+    },
+    {
+      name: 'base.palette.blackberry.200',
+      old: '#c3c2ff',
+      new: 'base.palette.indigo.200',
+    },
+    {
+      name: 'base.palette.blackberry.300',
+      old: '#8483e6',
+      new: 'base.palette.indigo.400',
+    },
+    {
+      name: 'base.palette.blackberry.400',
+      old: '#5c59e6',
+      new: 'base.palette.indigo.500',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.blackberry.500',
+      old: '#413fcc',
+      new: 'base.palette.indigo.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.blackberry.600',
+      old: '#2e2d91',
+      new: 'base.palette.indigo.900',
+      withWhiteText: true,
+    },
+  ],
+  'island-punch': [
+    {
+      name: 'base.palette.island-punch.100',
+      old: '#f5f0ff',
+      new: 'base.palette.purple.25',
+    },
+    {
+      name: 'base.palette.island-punch.200',
+      old: '#d2befa',
+      new: 'base.palette.purple.200',
+    },
+    {
+      name: 'base.palette.island-punch.300',
+      old: '#a88ae6',
+      new: 'base.palette.indigo.400',
+    },
+    {
+      name: 'base.palette.island-punch.400',
+      old: '#8660d1',
+      new: 'base.palette.purple.500',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.island-punch.500',
+      old: '#6345a1',
+      new: 'base.palette.purple.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.island-punch.600',
+      old: '#503882',
+      new: 'base.palette.purple.800',
+      withWhiteText: true,
+    },
+  ],
+  'grape-soda': [
+    {
+      name: 'base.palette.grape-soda.100',
+      old: '#feebff',
+      new: 'base.palette.magenta.50',
+    },
+    {
+      name: 'base.palette.grape-soda.200',
+      old: '#fac0ff',
+      new: 'base.palette.magenta.200',
+    },
+    {
+      name: 'base.palette.grape-soda.300',
+      old: '#de8ae6',
+      new: 'base.palette.purple.400',
+    },
+    {
+      name: 'base.palette.grape-soda.400',
+      old: '#c860d1',
+      new: 'base.palette.purple.500',
+    },
+    {
+      name: 'base.palette.grape-soda.500',
+      old: '#97499e',
+      new: 'base.palette.purple.600',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.grape-soda.600',
+      old: '#7C3882',
+      new: 'base.palette.purple.800',
+      withWhiteText: true,
+    },
+  ],
+  pomegranate: [
+    {
+      name: 'base.palette.pomegranate.100',
+      old: '#ffebf3',
+      new: 'base.palette.magenta.50',
+    },
+    {
+      name: 'base.palette.pomegranate.200',
+      old: '#ffbdd6',
+      new: 'base.palette.magenta.100',
+    },
+    {
+      name: 'base.palette.pomegranate.300',
+      old: '#ff5c9a',
+      new: 'base.palette.magenta.500',
+    },
+    {
+      name: 'base.palette.pomegranate.400',
+      old: '#f31167',
+      new: 'base.palette.red.400',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.pomegranate.500',
+      old: '#c70550',
+      new: 'base.palette.red.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.pomegranate.600',
+      old: '#99003a',
+      new: 'base.palette.red.800',
+      withWhiteText: true,
+    },
+  ],
+  'fruit-punch': [
+    {
+      name: 'base.palette.fruit-punch.100',
+      old: '#FFEEEE',
+      new: 'base.palette.red.25',
+    },
+    {
+      name: 'base.palette.fruit-punch.200',
+      old: '#ffbdbd',
+      new: 'base.palette.red.200',
+    },
+    {
+      name: 'base.palette.fruit-punch.300',
+      old: '#FF7E7E',
+      new: 'base.palette.red.300',
+    },
+    {
+      name: 'base.palette.fruit-punch.400',
+      old: '#ff4c4c',
+      new: 'base.palette.red.400',
+    },
+    {
+      name: 'base.palette.fruit-punch.500',
+      old: '#e12f2f',
+      new: 'base.palette.red.400',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.fruit-punch.600',
+      old: '#b82828',
+      new: 'base.palette.red.700',
+      withWhiteText: true,
+    },
+  ],
+  'root-beer': [
+    {
+      name: 'base.palette.root-beer.100',
+      old: '#faf3f0',
+      new: 'base.palette.coral.25',
+    },
+    {
+      name: 'base.palette.root-beer.200',
+      old: '#EBD7CF',
+      new: 'base.palette.coral.100',
+    },
+    {
+      name: 'base.palette.root-beer.300',
+      old: '#dcbbad',
+      new: 'base.palette.coral.200',
+    },
+    {
+      name: 'base.palette.root-beer.400',
+      old: '#ba9a8c',
+      new: 'base.palette.coral.200',
+    },
+    {
+      name: 'base.palette.root-beer.500',
+      old: '#8C7266',
+      new: 'base.palette.coral.200',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.root-beer.600',
+      old: '#664d42',
+      new: 'base.palette.amber.950',
+      withWhiteText: true,
+    },
+  ],
+  'toasted-marshmallow': [
+    {
+      name: 'base.palette.toasted-marshmallow.100',
+      old: '#fdf6e6',
+      new: 'base.palette.amber.25',
+    },
+    {
+      name: 'base.palette.toasted-marshmallow.200',
+      old: '#ebd6a9',
+      new: 'base.palette.orange.100',
+    },
+    {
+      name: 'base.palette.toasted-marshmallow.300',
+      old: '#e6bf6c',
+      new: 'base.palette.orange.200',
+    },
+    {
+      name: 'base.palette.toasted-marshmallow.400',
+      old: '#CC9E3B',
+      new: 'base.palette.orange.300',
+    },
+    {
+      name: 'base.palette.toasted-marshmallow.500',
+      old: '#b37f10',
+      new: 'base.palette.amber.500',
+    },
+    {
+      name: 'base.palette.toasted-marshmallow.600',
+      old: '#8C6000',
+      new: 'base.palette.amber.600',
+      withWhiteText: true,
+    },
+  ],
+  licorice: [
+    {
+      name: 'base.palette.licorice.100',
+      old: '#A1AAB3',
+      new: 'base.palette.slate.400',
+      systemTokens: [
+        'sys.color.bg.muted.softer',
+        'sys.color.fg.disabled',
+        'sys.color.border.input.disabled',
+      ],
+    },
+    {
+      name: 'base.palette.licorice.200',
+      old: '#7b858f',
+      new: 'base.palette.slate.500',
+      systemTokens: [
+        'sys.color.bg.muted.soft',
+        'sys.color.fg.muted.soft',
+        'sys.color.border.input.default',
+      ],
+    },
+    {
+      name: 'base.palette.licorice.300',
+      old: '#5E6A75',
+      new: 'base.palette.slate.600',
+      systemTokens: [
+        'sys.color.bg.muted.default',
+        'sys.color.fg.muted.default',
+        'sys.color.text.hint',
+      ],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.licorice.400',
+      old: '#4a5561',
+      new: 'base.palette.slate.700',
+      systemTokens: ['sys.color.fg.muted.strong'],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.licorice.500',
+      old: '#333d47',
+      new: 'base.palette.slate.800',
+      systemTokens: [
+        'sys.color.bg.muted.strong',
+        'sys.color.fg.muted.stronger',
+        'sys.color.border.input.strong',
+      ],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.licorice.600',
+      old: '#1f262e',
+      new: 'base.palette.slate.900',
+      withWhiteText: true,
+    },
+  ],
+  soap: [
+    {
+      name: 'base.palette.soap.100',
+      old: '#f6f7f8',
+      new: 'base.palette.slate.25',
+      systemTokens: ['sys.color.bg.alt.softer'],
+    },
+    {
+      name: 'base.palette.soap.200',
+      old: '#F0F1F2',
+      new: 'base.palette.slate.50',
+      systemTokens: ['sys.color.bg.alt.soft'],
+    },
+    {
+      name: 'base.palette.soap.300',
+      old: '#e8ebed',
+      new: 'base.palette.slate.100',
+      systemTokens: ['sys.color.bg.alt.default', 'sys.color.border.input.inverse'],
+    },
+    {
+      name: 'base.palette.soap.400',
+      old: '#DFE2E6',
+      new: 'base.palette.slate.200',
+      systemTokens: ['sys.color.bg.alt.strong', 'sys.color.border.divider'],
+    },
+    {
+      name: 'base.palette.soap.500',
+      old: '#ced3d9',
+      new: 'base.palette.slate.300',
+      systemTokens: ['sys.color.bg.alt.stronger', 'sys.color.border.container'],
+    },
+    {
+      name: 'base.palette.soap.600',
+      old: '#B9C0C7',
+      new: 'base.palette.slate.600',
+    },
+  ],
+  'french-vanilla': [
+    {
+      name: 'base.palette.french-vanilla.100',
+      old: '#ffffff',
+      new: 'base.palette.neutral.0',
+      systemTokens: ['sys.color.bg.default', 'sys.color.fg.default', 'sys.color.border.default'],
+    },
+    {
+      name: 'base.palette.french-vanilla.200',
+      old: '#ebebeb',
+      new: 'base.palette.neutral.50',
+    },
+    {
+      name: 'base.palette.french-vanilla.300',
+      old: '#d4d4d4',
+      new: 'base.palette.neutral.100',
+    },
+    {
+      name: 'base.palette.french-vanilla.400',
+      old: '#bdbdbd',
+      new: 'base.palette.neutral.200',
+    },
+    {
+      name: 'base.palette.french-vanilla.500',
+      old: '#a6a6a6',
+      new: 'base.palette.neutral.300',
+    },
+    {
+      name: 'base.palette.french-vanilla.600',
+      old: '#8f8f8f',
+      new: 'base.palette.neutral.400',
+    },
+  ],
+  'black-pepper': [
+    {
+      name: 'base.palette.black-pepper.100',
+      old: '#787878',
+      new: 'base.palette.neutral.500',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.black-pepper.200',
+      old: '#616161',
+      new: 'base.palette.neutral.700',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.black-pepper.300',
+      old: '#494949',
+      new: 'base.palette.neutral.900',
+      systemTokens: ['sys.color.fg.default'],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.black-pepper.400',
+      old: '#333333',
+      new: 'base.palette.neutral.950',
+      systemTokens: [
+        'sys.color.bg.contrast.default',
+        'sys.color.fg.strong',
+        'sys.color.border.contrast.default',
+      ],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.black-pepper.500',
+      old: '#1e1e1e',
+      new: 'base.palette.neutral.975',
+      systemTokens: [
+        'sys.color.bg.contrast.strong',
+        'sys.color.fg.stronger',
+        'sys.color.border.contrast.strong',
+      ],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.black-pepper.600',
+      old: '#000000',
+      new: 'base.palette.neutral.1000',
+      withWhiteText: true,
+    },
+  ],
+  coconut: [
+    {
+      name: 'base.palette.coconut.100',
+      old: '#F0EEEE',
+      new: '',
+    },
+    {
+      name: 'base.palette.coconut.200',
+      old: '#e3dfdf',
+      new: '',
+    },
+    {
+      name: 'base.palette.coconut.300',
+      old: '#d1cbcc',
+      new: '',
+    },
+    {
+      name: 'base.palette.coconut.400',
+      old: '#b3acac',
+      new: '',
+    },
+    {
+      name: 'base.palette.coconut.500',
+      old: '#9e9595',
+      new: '',
+    },
+    {
+      name: 'base.palette.coconut.600',
+      old: '#8F8687',
+      new: '',
+    },
+  ],
+  cappuccino: [
+    {
+      name: 'base.palette.cappuccino.100',
+      old: '#7A7374',
+      new: '',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.cappuccino.200',
+      old: '#706869',
+      new: '',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.cappuccino.300',
+      old: '#5E5757',
+      new: '',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.cappuccino.400',
+      old: '#4A4242',
+      new: '',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.cappuccino.500',
+      old: '#352f2f',
+      new: '',
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.cappuccino.600',
+      old: '#231f20',
+      new: '',
+      withWhiteText: true,
+    },
+  ],
+  'dragon-fruit': [
+    {
+      name: 'base.palette.dragon-fruit.100',
+      old: '#FBF1FF',
+      new: 'base.palette.purple.25',
+    },
+    {
+      name: 'base.palette.dragon-fruit.200',
+      old: '#EFD3FF',
+      new: 'base.palette.purple.100',
+    },
+    {
+      name: 'base.palette.dragon-fruit.300',
+      old: '#BE61F6',
+      new: 'base.palette.indigo.500',
+      systemTokens: ['sys.color.bg.ai.default'],
+    },
+    {
+      name: 'base.palette.dragon-fruit.400',
+      old: '#8C17D2',
+      new: 'base.palette.indigo.600',
+      systemTokens: ['sys.color.bg.ai.strong'],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.dragon-fruit.500',
+      old: '#6B11A3',
+      new: 'base.palette.indigo.700',
+      systemTokens: ['sys.color.bg.ai.stronger'],
+      withWhiteText: true,
+    },
+    {
+      name: 'base.palette.dragon-fruit.600',
+      old: '#4A0D71',
+      new: 'base.palette.indigo.900',
+      systemTokens: ['sys.color.bg.ai.strongest', 'sys.color.border.ai', 'sys.color.text.ai'],
+      withWhiteText: true,
+    },
+  ],
+};
