+++ packages/canvas-tokens-docs/stories/base/visual-testing/DeprecatedBase.stories.tsx
@@ -2,7 +2,7 @@ import {DeprecatedColorTokens} from '../examples/Color';
 
 export default {
   component: DeprecatedColorTokens,
-  title: 'Docs/Old Base Tokens (deprecated)',
+  title: 'Visual Tests/Old Base Tokens (deprecated)',
 };
 
 export const DeprecatedColors = {
