+++ packages/canvas-tokens-docs/stories/migration/examples/StaticUpdates.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {updatedSystemTokens} from './data/deprecatedMapping';
 import {TokenGrid} from '../../../components/TokenGrid';
-import {Swatch} from './DeprecatedTokens';
+import {SwatchWithText} from '../../../components/SwatchWithText';
 
 export const StaticUpdates = () => {
   const {otherColors} = updatedSystemTokens;
@@ -18,8 +18,12 @@ export const StaticUpdates = () => {
           {(token: any) => (
             <>
               <TokenGrid.RowItem>{token.name}</TokenGrid.RowItem>
-              <Swatch color={token.oldColor} name={token.old} withWhiteText={token.withWhiteText} />
-              <Swatch color={token.new} withWhiteText={token.withWhiteText} />
+              <SwatchWithText
+                color={token.oldColor}
+                name={token.old}
+                withWhiteText={token.withWhiteText}
+              />
+              <SwatchWithText color={token.new} withWhiteText={token.withWhiteText} />
               <TokenGrid.RowItem>
                 {!token.old ? '➕Added' : !token.new ? '❌ Removed' : '✅ Updated'}
               </TokenGrid.RowItem>
