+++ packages/canvas-tokens-docs/stories/migration/v3MigrationDeprecated.stories.mdx
@@ -1,6 +1,6 @@
 import {Markdown, Meta, Unstyled} from '@storybook/blocks';
 import {NavCard} from '../../components/NavCard';
-import {DeprecatedTokens} from './examples/DeprecatedTokens';
+import {DeprecatedTokens} from '../base/examples/DeprecatedTokens';
 
 <Meta title="Guides/Upgrade Guides/v3/Visual Changes/Base Deprecated Palette" />
